+++ pyKirara/__init__.py
@@ -1,2 +1,3 @@
 from .client import Kirara, KiraraException, Gacha, Info
 from .idol import Idol
+from .card import Card